rework rendering
Store tiles column-major order for easier single-column rendering for
scrolling, and rework the rendering so the tile logic can be shared between
full tile and single-column rendering.  While we are at it, add the sprite
(player) overlay support when rendering full tiles.

Also add tiles for the 'boy' player, and all 4 'girl' tiles.