Rework build system
arduino-mk assumes that the project makefile is called 'Makefile', so it
gets confused about the Makefile for native building.

To fix this, rework the build system so native/arduino builds can both be
built from 'Makefile' based on TARGET=native|arduino.

This also allows us to move common rules (E.G.  the gfx handling) to the
toplevel Makefile.  While we are at it, define the serial port to use for
arduino uploads.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>